requirements.txt
# Core ML and Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=0.19.0

transformers>=4.21.0
torch>=1.12.0
huggingface_hub>=0.16.0
sentencepiece>=0.1.99
safetensors>=0.3.0
nltk>=3.7
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0

# OpenAI for advanced imputation (if needed)
openai>=0.27.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.0.0
black>=22.0.0
flake8>=5.0.0

# API and Web (for potential deployment)
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=2.0.0,<2.11.2

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0